refactor(utils): move `backtickQuote()` to `utils/markdown`
The function had no `data-model` dependency and four of the five
packages that want it cannot reach `data-model` -- `utils` sits below it,
so the arrow can only point the other way. `packages/utils/src/markdown.ts`
is where every package can see it.

`content-hash` is the package that was stuck: its `Unknown encoding`
message interpolates a caller-supplied encoding name, which can hold a
backtick, and it had no way to quote one safely. Both of its sites now
do, with a test driving `` a`b `` through `digest()` and pinning the
doubled delimiter the result needs.

The round-trip tests move with the function, and the Markdown-derived
code-span reader they use moves with them. `data-model`'s
`message-quoting.test.ts` no longer needs that reader: it asserts each
message contains `backtickQuote(input)`, which is the property that test
is actually about -- that the call site quotes at all. Whether the
quoting is correct is pinned once, in `utils`, against a reader written
from the specification.
