+++ packages/data-model/src/codec-common/EmptyReconstructionContext.ts
@@ -5,10 +5,10 @@
  * structurally flat).
  */
 
+import { backtickQuote } from "@commonfabric/utils/markdown";
 import type { FabricInstance } from "@/interface.ts";
 import type { ReconstructionContext } from "./interface.ts";
 import { BaseReconstructionContext } from "./BaseReconstructionContext.ts";
-import { backtickQuote } from "@/value-debug.ts";
 
 /**
  * `ReconstructionContext` whose `getCell()` always throws. `shouldDeepFreeze`
