+++ packages/data-model/src/fabric-instances/FabricNativeWrapper.ts
@@ -1,6 +1,6 @@
+import { backtickQuote } from "@commonfabric/utils/markdown";
 import type { FabricInstance } from "@/interface.ts";
 import { BaseFabricInstance, DEEP_CLONE_CORE } from "./BaseFabricInstance.ts";
-import { backtickQuote } from "@/value-debug.ts";
 
 /**
  * Abstract base class for `FabricInstance` wrappers that bridge native JS
