+++ packages/utils/deno.jsonc
@@ -28,6 +28,7 @@
     "./env": "./src/env.ts",
     "./equal-ignoring-symbols": "./src/equal-ignoring-symbols.ts",
     "./logger": "./src/logger.ts",
+    "./markdown": "./src/markdown.ts",
     "./objects": "./src/objects.ts",
     "./path-key-map": "./src/path-key-map.ts",
     "./sandbox-contract": "./src/sandbox-contract.ts",
